// apps/~internal/_generated/theme-mappings.ts
/* file generated by apps/theme/export-theme.ts */

export type TailwindThemeMapping = [
  name: string,
  tailwindVar: string,
  type: 'color' | 'px',
]

export const tailwindThemeMappings: TailwindThemeMapping[] = [
  ['accent', '--color-th_accent', 'color'],
  ['focus', '--color-th_focus', 'color'],
  ['link', '--color-th_link', 'color'],
  ['separator', '--color-th_separator', 'color'],
  ['baseBackground', '--background-color-th_base', 'color'],
  ['baseBorder', '--border-color-th_base', 'color'],
  ['baseContent', '--text-color-th_base', 'color'],
  ['baseContentDimmed', '--text-color-th_base-dimmed', 'color'],
  ['baseHoveredBackground', '--background-color-th_base-hovered', 'color'],
  ['frameBackground', '--background-color-th_frame', 'color'],
  ['frameContent', '--text-color-th_frame', 'color'],
  ['frameBorder', '--border-color-th_frame', 'color'],
  ['frameRadius', '--radius-th_frame', 'px'],
  ['iconInfoBackground', '--background-color-th_icon-info', 'color'],
  ['iconInfoContent', '--text-color-th_icon-info', 'color'],
  ['primaryBackground', '--background-color-th_primary', 'color'],
  ['primaryBorder', '--border-color-th_primary', 'color'],
  ['primaryContent', '--text-color-th_primary', 'color'],
  [
    'primaryHoveredBackground',
    '--background-color-th_primary-hovered',
    'color',
  ],
  ['primaryHoveredBorder', '--border-color-th_primary-hovered', 'color'],
  ['negativeBackground', '--background-color-th_negative', 'color'],
  ['negativeContent', '--text-color-th_negative', 'color'],
  ['positiveBackground', '--background-color-th_positive', 'color'],
  ['positiveContent', '--text-color-th_positive', 'color'],
  ['fieldBackground', '--background-color-th_field', 'color'],
  ['fieldBorder', '--border-color-th_field', 'color'],
  ['fieldContent', '--text-color-th_field', 'color'],
  ['fieldErrorBorder', '--border-color-th_field-error', 'color'],
  ['fieldFocusedBackground', '--background-color-th_field-focused', 'color'],
  ['fieldFocusedContent', '--text-color-th_field-focused', 'color'],
]